value calculation

+++ src/pentago/game.rs
@@ -3,14 +3,14 @@ use pentago::board;
 use pentago::board::Board;
 
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct Game {
-    pub victory: i32,
+    pub victory: u8,
     pub board: Board
 }
 
 impl Game {
-    pub fn new(dim: i32, length: i32, victory: i32) -> Game {
+    pub fn new(dim: u8, length: u8, victory: u8) -> Game {
         Game {
             victory: victory,
             board: Board::new(dim, length)
